add check for number of token pair auctions on DutchExchange
if there is not enough historical data (currentAuctionIndex does not exceed NUM_PRICE_POINTS) then the FutarchyChallengeFactory.createChallenge will revert when it tries to determine the upper/lower bounds for the scalar markets. This adds a check for this condition and throws an error.

Since this required us to access an instance of FutarchyChallengeFactory, which stores the LMSRMarketMaker address, the LMSRMarketMaker config was removed and replaced with a call to FutarchyChallengeFactory.